+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/ClubNewsActivity.java
@@ -364,6 +364,7 @@ public class ClubNewsActivity extends BaseActivity {
                 //you can handle the errors here
                 Log.e(LOG_TAG, "RetrofitError : " + error);
                 hideProgress();
+                reportRollBarException(ClubNewsActivity.class.getSimpleName(), error.toString());
                 showAlertOk("" + getResources().getString(R.string.error_server_problem), false);
             }
         });
@@ -396,6 +397,7 @@ public class ClubNewsActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(ClubNewsActivity.class.getSimpleName(), e.toString());
         }
 
         //Dismiss progress dialog.
@@ -529,6 +531,7 @@ public class ClubNewsActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(ClubNewsActivity.class.getSimpleName(), e.toString());
         }
 
         //Dismiss progress dialog.