+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/ClubNewsDetailActivity.java
@@ -363,6 +363,7 @@ public class ClubNewsDetailActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(ClubNewsDetailActivity.class.getSimpleName(), e.toString());
         }
 
         //Dismiss progress dialog.
@@ -399,6 +400,7 @@ public class ClubNewsDetailActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(ClubNewsDetailActivity.class.getSimpleName(), e.toString());
         }
 
         //resetValues();