+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/CompetitionDetailActivity.java
@@ -410,6 +410,7 @@ public class CompetitionDetailActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(CompetitionDetailActivity.class.getSimpleName(), e.toString());
         }
         hideProgress();
     }
@@ -598,6 +599,7 @@ public class CompetitionDetailActivity extends BaseActivity {
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
             e.printStackTrace();
+            reportRollBarException(CompetitionDetailActivity.class.getSimpleName(), e.toString());
         }
         hideProgress();
     }