+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/CompetitionsActivity.java
@@ -245,17 +245,26 @@ public class CompetitionsActivity extends BaseActivity {
                             isCurrent = false;
                             break;
                     }
-
                     resetArrayData();
 
                     if (iPopItemPos < 2) {
                         resetCalendar();
-                        createDateForData();
+                        //createDateForData();
                     } else {
+                        strDate = "01";
+                        iMonth = iCurrentMonth;
+                        iYear = iCurrentYear;
+                        //createDateForData();
+
                         //Set ENABLE/DISABLE state of ICONS on change tab or pressed.
                         //resetMonthsNavigationIcons();
-                        callCompetitionsWebService();
+                        //callCompetitionsWebService();
                     }
+
+
+                    mCalendarInstance = new GregorianCalendar(iYear, (iMonth - 1), Integer.parseInt(strDate));
+                    iNumOfDays = mCalendarInstance.getActualMaximum(Calendar.DAY_OF_MONTH);
+                    createDateForData();
                     return true;
                 }
             };
@@ -640,7 +649,7 @@ public class CompetitionsActivity extends BaseActivity {
             }
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            reportRollBarException(CompetitionsActivity.class.getSimpleName(), e.toString());
         }
 
         //Update Month title name.