+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/EligiblePlayersActivity.java
@@ -466,7 +466,7 @@ public class EligiblePlayersActivity extends BaseActivity {
             // hideProgress();
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            reportRollBarException(EligiblePlayersActivity.class.getSimpleName(), e.toString());
         }
     }