+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/activites/YourAccountActivity.java
@@ -77,6 +77,8 @@ public class YourAccountActivity extends BaseActivity {
 
      /* -- INTERNET CONNECTION PARAMETERS -- */
 
+    public static boolean isRefreshEnable;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -84,6 +86,8 @@ public class YourAccountActivity extends BaseActivity {
 
         //Initialize view resources.
         ButterKnife.bind(this);
+
+        isRefreshEnable = true;
 
         if (tbMyAccount != null) {
             setSupportActionBar(tbMyAccount);
@@ -116,8 +120,9 @@ public class YourAccountActivity extends BaseActivity {
     @Override
     protected void onResume() {
         super.onResume();
-
-        updateFragment(new MyAccountTabFragment(getiOpenTabPosition()));
+        if (isRefreshEnable) {
+            updateFragment(new MyAccountTabFragment(getiOpenTabPosition()));
+        }
     }
 
     @Override