+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/ui/fragments/MembersFragment.java
@@ -17,6 +17,7 @@ import android.widget.AdapterView;
 import android.widget.TextView;
 
 import com.google.gson.JsonObject;
+import com.mh.systems.hartsbourne.ui.activites.YourAccountActivity;
 import com.newrelic.com.google.gson.reflect.TypeToken;
 import com.mh.systems.hartsbourne.R;
 import com.mh.systems.hartsbourne.ui.activites.BaseActivity;
@@ -356,7 +357,7 @@ public class MembersFragment extends Fragment {
             }
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            ((YourAccountActivity) getActivity()).reportRollBarException(MembersFragment.class.getSimpleName(), e.toString());
         }
 
         //Dismiss progress dialog.