+++ Teeside/MHSystems/app/src/main/java/com/mh/systems/teesside/ui/fragments/FinanceFragment.java
@@ -1,7 +1,6 @@
 package com.mh.systems.teesside.ui.fragments;
 
 
-import android.content.Context;
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.os.Bundle;
@@ -9,15 +8,12 @@ import android.support.v4.app.Fragment;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.OrientationHelper;
-import android.support.v7.widget.PopupMenu;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
@@ -37,11 +33,6 @@ import com.mh.systems.teesside.web.models.FinanceAPI;
 import com.mh.systems.teesside.web.models.FinanceResultItems;
 import com.mh.systems.teesside.web.models.TransactionListData;
 import com.mh.systems.teesside.web.models.finance.FinanceFilter;
-import com.mh.systems.teesside.web.models.pursebalance.AJsonParamsPurseApi;
-import com.mh.systems.teesside.web.models.pursebalance.AccountBalance;
-import com.mh.systems.teesside.web.models.pursebalance.PurseBalanceApi;
-import com.mh.systems.teesside.web.models.pursebalance.PurseBalanceResponse;
-import com.newrelic.com.google.gson.Gson;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
@@ -51,7 +42,6 @@ import retrofit.Callback;
 import retrofit.RestAdapter;
 import retrofit.RetrofitError;
 
-
 /**
  * Created by karan@ucreate.co.in to load and FINANCE
  * tab content on 12/23/2015.
@@ -60,10 +50,8 @@ public class FinanceFragment extends Fragment {
     /*********************************
      * INSTANCES OF LOCAL DATA TYPE
      *******************************/
-    public static final String LOG_TAG = FinanceFragment.class.getSimpleName();
-    ArrayList<TransactionListData> transactionListDataArrayList = new ArrayList<>();
-
-    ArrayList<AccountBalance> purseBalanceDatas = new ArrayList<>();
+    public final String LOG_TAG = FinanceFragment.class.getSimpleName();
+   // ArrayList<TransactionListData> transactionListDataArrayList = new ArrayList<>();
 
     /*++ Filter type for Today, 1 Week, 1 Month, 3 Months, 6 Months, 1 Year or From Start++*/
     int iFilterType = 2; //By Default 1 month will be selected.
@@ -80,13 +68,17 @@ public class FinanceFragment extends Fragment {
     View viewRootFragment;
     TextView tvLabelCardBalance, tvCardBalance, tvDateHeading;
     TextView tvLabelYourInvoice, tvYourInvoice;
+    TextView tvAmount, tvBalance;
     ImageView ivFilter;
-    LinearLayout llTransactionUI;
-    View vwPopMenu;
     Intent intent;
     Button btTopUp;
 
     Typeface tpRobotoMedium, tpRobotoRegular;
+
+    //Instance of Transaction listview.
+    //ListView lvTransactionList;
+
+    //FinanceAdapter financeAdapter;
 
     //List of type books this list will store type Book which is our data model
     private FinanceAPI financeAPI;
@@ -96,19 +88,6 @@ public class FinanceFragment extends Fragment {
     FinanceResultItems financeResultItems;
 
     LinearLayout llFinanceGroup;
-    LinearLayout llPurseType;
-    FrameLayout flPurseGroup, flInvoicesGroup, flTopUpGroup;
-
-    //Pop Menu to show Categories of Course Diary.
-    PopupMenu popupMenu;
-
-    MenuItem financeMenuItem = null;
-
-    PurseBalanceApi purseBalanceApi;
-    AJsonParamsPurseApi aJsonParamsPurseApi;
-
-    private PurseBalanceResponse mPurseBalanceResponse;
-    private Context mContext;
 
     ArrayList<FinanceFilter> financeFilterArrayList = new ArrayList<>();
     List<TransactionListData> TransactionList = new ArrayList<>();
@@ -121,9 +100,10 @@ public class FinanceFragment extends Fragment {
 
         initViewResources();
 
-        initPurseCategory();
-
         setFontTypeface();
+
+//        financeAdapter = new FinanceAdapter(getActivity(), transactionListDataArrayList);
+//        lvTransactionList.setAdapter(financeAdapter);
 
         financeRecycleAdapter = new FinanceRecycleAdapter(financeFilterArrayList, getActivity());
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity(), OrientationHelper.VERTICAL, false);
@@ -132,14 +112,7 @@ public class FinanceFragment extends Fragment {
         rvTransactionList.setItemAnimator(new DefaultItemAnimator());
         rvTransactionList.setAdapter(financeRecycleAdapter);
 
-        llPurseType.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                popupMenu.show();
-            }
-        });
-        popupMenu.setOnMenuItemClickListener(mPurseCategoryListener);
-
+        //lvTransactionList.setOnItemClickListener(mFinanceListListener);
         btTopUp.setOnClickListener(mTopUpListener);
 
         return viewRootFragment;
@@ -149,20 +122,13 @@ public class FinanceFragment extends Fragment {
     public void setUserVisibleHint(boolean isVisibleToUser) {
         super.setUserVisibleHint(isVisibleToUser);
 
-        if (isVisibleToUser && ((YourAccountActivity)getActivity()).getiBalanceType() == 0) {
+        if (isVisibleToUser) {
             callFinanceWebService();
-            ((YourAccountActivity) mContext).updateFilterIcon(0);
-            ((YourAccountActivity) mContext).setiOpenTabPosition(2);
+            ((YourAccountActivity) getActivity()).updateFilterIcon(0);
+            ((YourAccountActivity) getActivity()).setiOpenTabPosition(2);
 
             YourAccountActivity.isRefreshEnable = false;
         }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-
-        mContext = context;
     }
 
     /**
@@ -173,46 +139,36 @@ public class FinanceFragment extends Fragment {
         callFinanceWebService();
     }
 
-    private View.OnClickListener mTopUpListener = new View.OnClickListener() {
+
+//    public AdapterView.OnItemClickListener mFinanceListListener = new AdapterView.OnItemClickListener() {
+//        @Override
+//        public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+//
+//            intent = new Intent(getActivity(), FinanceDetailWebActivity.class);
+//            intent.putExtra("IsTopup", transactionListDataArrayList.get(position).getIsTopup());
+//            intent.putExtra("iTransactionId", transactionListDataArrayList.get(position).getTransactionId());
+//            intent.putExtra("strMemberId", ((YourAccountActivity) getActivity()).getMemberId());
+//            intent.putExtra("titleOfScreen", transactionListDataArrayList.get(position).getTitle());
+//            startActivity(intent);
+//        }
+//    };
+
+    public View.OnClickListener mTopUpListener = new View.OnClickListener() {
         @Override
         public void onClick(View v) {
 
             if (financeResultItems != null) {
-                intent = new Intent(mContext, TopUpActivity.class);
+                intent = new Intent(getActivity(), TopUpActivity.class);
                 intent.putExtra("strClosingBalance", strClosingBalance);
                 startActivity(intent);
             }
         }
     };
 
-    private PopupMenu.OnMenuItemClickListener mPurseCategoryListener = new PopupMenu.OnMenuItemClickListener() {
-        @Override
-        public boolean onMenuItemClick(MenuItem item) {
-
-            financeMenuItem = item;
-
-            if (((BaseActivity) mContext).isOnline(mContext)) {
-                requestPurseApiService();
-            }
-
-            return true;
-        }
-    };
-
     /**
-     * Implements this method to initialized the purse
-     * category of finance.
+     * Implements this method to initialize all
+     * view resources.
      */
-    private void initPurseCategory() {
-
-        popupMenu = new PopupMenu(mContext, vwPopMenu);
-        popupMenu.inflate(R.menu.menu_finance_purse);
-
-        //Initially display title at 0 index.
-        tvLabelCardBalance.setText(popupMenu.getMenu().getItem(0) + " " + getString(R.string.title_text_balance));
-    }
-
-
     private void initViewResources() {
         tvCardBalance = (TextView) viewRootFragment.findViewById(R.id.tvCardBalance);
         tvDateHeading = (TextView) viewRootFragment.findViewById(R.id.tvDateHeading);
@@ -221,19 +177,14 @@ public class FinanceFragment extends Fragment {
         tvLabelYourInvoice = (TextView) viewRootFragment.findViewById(R.id.tvLabelYourInvoice);
         tvYourInvoice = (TextView) viewRootFragment.findViewById(R.id.tvYourInvoice);
 
+        tvAmount = (TextView) viewRootFragment.findViewById(R.id.tvAmount);
+        tvBalance = (TextView) viewRootFragment.findViewById(R.id.tvBalance);
+
         ivFilter = (ImageView) viewRootFragment.findViewById(R.id.ivFilter);
-        vwPopMenu = (View) viewRootFragment.findViewById(R.id.vwPopMenu);
 
         rvTransactionList = (RecyclerView) viewRootFragment.findViewById(R.id.rvTransactionList);
 
         llFinanceGroup = (LinearLayout) viewRootFragment.findViewById(R.id.llFinanceGroup);
-        llPurseType = (LinearLayout) viewRootFragment.findViewById(R.id.llPurseType);
-
-        llTransactionUI = (LinearLayout) viewRootFragment.findViewById(R.id.llTransactionUI);
-
-        flPurseGroup = (FrameLayout) viewRootFragment.findViewById(R.id.flPurseGroup);
-        flInvoicesGroup = (FrameLayout) viewRootFragment.findViewById(R.id.flInvoicesGroup);
-        flTopUpGroup = (FrameLayout) viewRootFragment.findViewById(R.id.flTopUpGroup);
 
         btTopUp = (Button) viewRootFragment.findViewById(R.id.btTopUp);
     }
@@ -246,7 +197,7 @@ public class FinanceFragment extends Fragment {
         /**
          *  Check internet connection before hitting server request.
          */
-        if (((BaseActivity) mContext).isOnline(mContext)) {
+        if (((BaseActivity) getActivity()).isOnline(getActivity())) {
             requestFinanceService();
         }
     }
@@ -257,14 +208,14 @@ public class FinanceFragment extends Fragment {
      */
     private void requestFinanceService() {
 
-        ((BaseActivity) mContext).showPleaseWait("Loading...");
+        ((BaseActivity) getActivity()).showPleaseWait("Loading...");
 
         myAccountJsonParams = new FinanceAJsonParams();
         myAccountJsonParams.setCallid(ApplicationGlobal.TAG_NEW_GCLUB_CALL_ID);
         myAccountJsonParams.setDateRange(iFilterType);
-        myAccountJsonParams.setMemberId(((YourAccountActivity) mContext).getMemberId());
+        myAccountJsonParams.setMemberId(((YourAccountActivity) getActivity()).getMemberId());
 
-        financeAPI = new FinanceAPI((((YourAccountActivity) mContext).getClientId()), "GetAccStatement", myAccountJsonParams, "TRANSACTION", ApplicationGlobal.TAG_GCLUB_MEMBERS);
+        financeAPI = new FinanceAPI((((YourAccountActivity) getActivity()).getClientId()), "GetAccStatement", myAccountJsonParams, "TRANSACTION", ApplicationGlobal.TAG_GCLUB_MEMBERS);
 
         //Creating a rest adapter
         RestAdapter adapter = new RestAdapter.Builder()
@@ -286,7 +237,7 @@ public class FinanceFragment extends Fragment {
             public void failure(RetrofitError error) {
                 //you can handle the errors here
                 Log.e(LOG_TAG, "RetrofitError : " + error);
-                ((BaseActivity) mContext).hideProgress();
+                ((BaseActivity) getActivity()).hideProgress();
             }
         });
     }
@@ -301,8 +252,10 @@ public class FinanceFragment extends Fragment {
 
         Type type = new TypeToken<FinanceResultItems>() {
         }.getType();
-        financeResultItems = new Gson().fromJson(jsonObject.toString(), type);
+        financeResultItems = new com.newrelic.com.google.gson.Gson().fromJson(jsonObject.toString(), type);
 
+        //Clear array list before inserting items.
+        //transactionListDataArrayList.clear();
         financeFilterArrayList.clear();
         TransactionList.clear();
 
@@ -324,6 +277,8 @@ public class FinanceFragment extends Fragment {
                     financeFilterArrayList.add(new FinanceFilter(FinanceFilter.TYPE_DATA, "", TransactionList.get(iCount)));
                 }
 
+                //transactionListDataArrayList.addAll(financeResultItems.getData().getTransactionList());
+
                 if (financeFilterArrayList.size() == 0) {
                     ((BaseActivity) getActivity()).showAlertMessage("No Transaction Found.");
                 }
@@ -341,7 +296,7 @@ public class FinanceFragment extends Fragment {
             }
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            ((BaseActivity)getActivity()).reportRollBarException(FinanceFragment.class.getSimpleName(), e.toString());
+            ((YourAccountActivity) getActivity()).reportRollBarException(FinanceFragment.class.getSimpleName(), e.toString());
         }
 
         //Dismiss progress dialog.
@@ -388,8 +343,8 @@ public class FinanceFragment extends Fragment {
      * Implements a method to set Font style.
      */
     private void setFontTypeface() {
-        tpRobotoRegular = Typeface.createFromAsset(mContext.getAssets(), "fonts/Roboto-Regular.ttf");
-        tpRobotoMedium = Typeface.createFromAsset(mContext.getAssets(), "fonts/Roboto-Medium.ttf");
+        tpRobotoRegular = Typeface.createFromAsset(getActivity().getAssets(), "fonts/Roboto-Regular.ttf");
+        tpRobotoMedium = Typeface.createFromAsset(getActivity().getAssets(), "fonts/Roboto-Medium.ttf");
 
         tvLabelCardBalance.setTypeface(tpRobotoRegular);
         tvLabelYourInvoice.setTypeface(tpRobotoRegular);
@@ -397,128 +352,7 @@ public class FinanceFragment extends Fragment {
         tvYourInvoice.setTypeface(tpRobotoMedium);
 
         tvDateHeading.setTypeface(tpRobotoMedium);
+        tvAmount.setTypeface(tpRobotoMedium);
+        tvBalance.setTypeface(tpRobotoMedium);
     }
-
-    /* ++++++++++++++++ START OF PURSE API FEATURE ++++++++++++++++ */
-
-    private void requestPurseApiService() {
-
-        ((BaseActivity) mContext).showPleaseWait("Loading...");
-
-        aJsonParamsPurseApi = new AJsonParamsPurseApi();
-        aJsonParamsPurseApi.setCallid(ApplicationGlobal.TAG_NEW_GCLUB_CALL_ID);
-        aJsonParamsPurseApi.setVersion(ApplicationGlobal.TAG_GCLUB_VERSION);
-        aJsonParamsPurseApi.setMemberId(((YourAccountActivity) mContext).getMemberId());
-
-        purseBalanceApi = new PurseBalanceApi((((YourAccountActivity) mContext).getClientId()), "GetMemberPurseBalances", aJsonParamsPurseApi, ApplicationGlobal.TAG_GCLUB_WEBSERVICES, ApplicationGlobal.TAG_GCLUB_MEMBERS);
-
-        //Creating a rest adapter
-        RestAdapter adapter = new RestAdapter.Builder()
-                .setEndpoint(WebAPI.API_BASE_URL)
-                .build();
-
-        //Creating an object of our api interface
-        WebServiceMethods api = adapter.create(WebServiceMethods.class);
-
-        //Defining the method
-        api.getFinancePurseBalance(purseBalanceApi, new Callback<JsonObject>() {
-            @Override
-            public void success(JsonObject jsonObject, retrofit.client.Response response) {
-
-                updatePurseApiSuccess(jsonObject);
-            }
-
-            @Override
-            public void failure(RetrofitError error) {
-                //you can handle the errors here
-                Log.e(LOG_TAG, "RetrofitError : " + error);
-                ((BaseActivity) mContext).hideProgress();
-            }
-        });
-    }
-
-    private void updatePurseApiSuccess(JsonObject jsonObject) {
-
-        Log.e(LOG_TAG, "SUCCESS RESULT : " + jsonObject.toString());
-
-        Type type = new TypeToken<PurseBalanceResponse>() {
-        }.getType();
-        mPurseBalanceResponse = new Gson().fromJson(jsonObject.toString(), type);
-
-        if (mPurseBalanceResponse.getData() != null) {
-            purseBalanceDatas.addAll(mPurseBalanceResponse.getData().getAccountBalances());
-        }
-
-        if (financeMenuItem != null) {
-            tvLabelCardBalance.setText(financeMenuItem.getTitle() + " " + getString(R.string.title_text_balance));
-            updatePurseType(financeMenuItem);
-        }
-
-        ((BaseActivity) mContext).hideProgress();
-    }
-
-    private String getPurseBalance(int iCrnID) {
-
-        for (int iCount = 0; iCount < purseBalanceDatas.size(); iCount++) {
-
-            if (purseBalanceDatas.get(iCount).getCrnID() == iCrnID) {
-
-                return purseBalanceDatas.get(iCount).getBalance();
-            }
-        }
-        return null;
-    }
-
-    private void updatePurseType(MenuItem menuItemInstance) {
-
-        switch (menuItemInstance.getItemId()) {
-            case R.id.item_general:
-                ((YourAccountActivity)mContext).setiBalanceType(0);
-                tvCardBalance.setText(getPurseBalance(0));
-                CollapsePurseUI();
-                break;
-
-            case R.id.item_Competitions:
-                ((YourAccountActivity)mContext).setiBalanceType(1);
-                ExpandPurseUI();
-                tvCardBalance.setText(getPurseBalance(1));
-                break;
-
-            case R.id.item_Associates:
-                ((YourAccountActivity)mContext).setiBalanceType(8);
-                ExpandPurseUI();
-                tvCardBalance.setText(getPurseBalance(8));
-                break;
-        }
-    }
-
-    /**
-     * Visible the following Purse Api according to position and hide
-     * others:
-     * - General
-     * - Competitions
-     * - Associate
-     */
-    private void ExpandPurseUI() {
-        flPurseGroup.setVisibility(View.VISIBLE);
-        flInvoicesGroup.setVisibility(View.GONE);
-        llTransactionUI.setVisibility(View.GONE);
-        flTopUpGroup.setVisibility(View.GONE);
-
-        //Hide Filter Icon on top right if General not selected.
-        ((YourAccountActivity) mContext).updateFilterIcon(View.GONE);
-    }
-
-    private void CollapsePurseUI() {
-        flPurseGroup.setVisibility(View.VISIBLE);
-        flInvoicesGroup.setVisibility(View.VISIBLE);
-        llTransactionUI.setVisibility(View.VISIBLE);
-        flTopUpGroup.setVisibility(View.VISIBLE);
-
-        //Show top right filter to choose transactions according to date.
-        ((YourAccountActivity) mContext).updateFilterIcon(View.VISIBLE);
-    }
-
-
-     /* ++++++++++++++++  END OF PURSE API FEATURE ++++++++++++++++ */
 }