+++ Teeside/MHSystems/app/src/main/java/com/mh/systems/teesside/ui/fragments/FriendsFragment.java
@@ -222,7 +222,7 @@ public class FriendsFragment extends Fragment {
             ((BaseActivity) getActivity()).hideProgress();
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            ((MembersActivity) getActivity()).reportRollBarException(ContactUsFragment.class.getSimpleName(), e.toString());
+            ((MembersActivity) getActivity()).reportRollBarException(FriendsFragment.class.getSimpleName(), e.toString());
         }
     }